MPEG4Extractor: Keep CTS values positive
Media files with zero STTS delta in video track exists, probably malformed files. If the same
video track has edit shift entry to shift(prepone) video frames, then CTS could get negative.
Since negative CTS values are not meaningful, this CL fixes that by only subtracting the lower
of CTS and edit shift value from CTS.

Bug: 117402420

Test: cts-tradefed run cts-dev -m CtsMediaTestCases --compatibility:module-arg CtsMediaTestCases:
      include-annotation:android.platform.test.annotations.RequiresDevice

+++ media/extractors/mp4/MPEG4Extractor.h
@@ -83,9 +83,13 @@ private:
         bool includes_expensive_metadata;
         bool skipTrack;
         bool has_elst;
+        /* signed int, ISO Spec allows media_time = -1 for other use cases.
+         * but we don't support empty edits for now.
+         */
         int64_t elst_media_time;
         uint64_t elst_segment_duration;
-        int32_t elstShiftStartTicks;
+        // unsigned int, shift start offset only when media_time > 0.
+        uint64_t elstShiftStartTicks;
         bool subsample_encryption;
     };
